+++ stellar/stellar_backend/controllers/notifications_controller.py
@@ -1,16 +1,29 @@
 from rest_framework.decorators import api_view
-from ..services import notification_service
+from rest_framework import status
+from rest_framework.response import Response
+
+from ..services import notification_service, session_service
+from ..static import constants as const
 
 @api_view(['GET'])
 def get_notifications(request, user_id):
+    if not session_service.is_token_valid(request.META):
+        return Response(const.INVALID_TOKEN, status = status.HTTP_401_UNAUTHORIZED)
+
     return notification_service.get_notifications(user_id)
 
 @api_view(['DELETE'])
 def delete_notification(request, notif_id):
+    if not session_service.is_token_valid(request.META):
+        return Response(const.INVALID_TOKEN, status = status.HTTP_401_UNAUTHORIZED)
+    
     return notification_service.delete_notification(notif_id)
 
 @api_view(['DELETE'])
 def delete_user_notifications(request, user_id):
+    if not session_service.is_token_valid(request.META):
+        return Response(const.INVALID_TOKEN, status = status.HTTP_401_UNAUTHORIZED)
+    
     return notification_service.delete_user_notifications(user_id)
 
 @api_view(['POST'])
